requirements.txt
colorama==0.4.6
python_docx==0.8.11
python_pptx==0.6.21
Requests==2.31.0
setuptools==65.6.3
tqdm==4.64.1
sphinx_rtd_theme==1.2.2
